Created an analyzer configuration to test vqwk channels in all the rocs.
Basically, each adc channel is analyzed as a bcm but only the Mps_Tree is saved in the rootfiles with the stem vqwk_check_.
I created a seperated channel map for this with each channel being named with its roc, module sn number and location in the module.
e.g. qwk_roc2sn13ch0 is the first channel in module 13 of roc2.

The script Extensions/Macros/vqwk_channel_check.C to can be used to open the rootfiles created by this configuration to plot the raw_hw_sum of the adc channels, in a given roc, after converting them to voltage using the conversion 76.20muV/sample (from VQWK manual). 

It will be a whole lot easier to check all the adc channels in the future using this configuration.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3409 b9700a98-9937-0410-b777-f318d6eb5bff