Made a copy of M. Dalton's MakeSlug.C script to work on a single run and create a "sluglet" file with the same leaves as the slugfiles produced previously. It uses the fix committed by J. Leacock into the pass3 copy of MakeSlugs that addresses a problem with older versions of ROOT having issues with friending in trees with no name.
The big change is that the script no longer requires a file with the runs+runlets defined in it, but rather takes in the run number as an argument and then parses 
the filenames of ROOTFiles in QW_ROOTFILES for the appropriate runs, and then extracts the segment number from the filename.
The rest of the script is identical to the MakeSlugs.C copy in the pass3 directory of Extentions/Macros/Parity/SlugPlots.

The script can be run by invoking the following	command:

root -l	'MakeSluglet.C+(RUNNUM,"full/path/to/leaflist.txt","full/path/to/outputFile.root","QwPass1")

Where the first	three arguments	are clear, but the 4th is whatever rootfile stem you want to be	looking	for.

NOTE THAT IT *DOES* LOOK FOR ROOTFILES WITH SUFFIXES OF THE FORMAT ".trees.root".

It does *NOT* currently friend in the regressed quantities. I need to figure out where the regressed ROOTfiles are being kept and talk to people about whether we want both
regressed and unregressed quantities or just regressed. (The latter would be very easy if I knew where the ROOTfiles were, I'm not too sure yet about how the script
would have to change to include both types)




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3418 b9700a98-9937-0410-b777-f318d6eb5bff