Moved the WienMode determination (Forward, Backward, HorizTransverse, VerticalTransverse) out of the Blinder class and into the EPICS class.  The Blinder now queries the EPICS class for the WienMode instead of for the variables which define it.
Added an "effective" EPICS variable to the EPICS variable list:
"WienMode".  This is determined internally by the EPICS class
for each EPICS event, and otherwise treated as if it were a
real EPICS variable.
It will be written into the slow_controls_strings table, if
we add it to the DB's variable list.

Added a NominalWienAngle setting in the EPICS map file; this is
the Wien setting which we would use for Horizontal-Only (or the
effective setting for Flip-Right) in order to give the electron
spin pointing forward in the H+ events for IHWP OUT.
This should be a value of +30 degrees for the Run1 data, and
may be something else for Run2.
Note that this could potentially be used to handle an odd-pi spin
precession, if we have one at 2-pass beam.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3427 b9700a98-9937-0410-b777-f318d6eb5bff