This is a significant change in the beamline classes. The scaler BCM and scaler clock signals are now handled by the BCM and Clock classes, so only actual halo monitors are handled by the QwHaloMonitor class.
The beamline mapping and pedestal files have been changed
in support of this.

The beamline subsystem uses the greedy parameter file functionality
and has had more code centralized, and the decoding of the channel
map file has more handling done by the data element classes and less
done directly in the subsystem classes.


This has been checked against trunk revision 3523 using run 12079.

NOTE:  the scaler values in the beamline system have changed slightly,
but that is because the time-normalization of the scalers had an error
in the previous trunk versions.
That error is resolved as a side-effect of the other changes; this was
not intended, but is understood, and this version is more correct than
what was in the trunk.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3524 b9700a98-9937-0410-b777-f318d6eb5bff