Change one of the functions named     QwSubsystemArrayParity::GetEventcutErrorFlag() to be named     QwSubsystemArrayParity::UpdateEventcutErrorFlag() instead.
This function was only used in QwEventRing to update
the ErrorFlag values after determining if the stability
cut had failed for any elements.

This was needed to ensure that only the const version
of GetEventcutErrorFlag existed, so that the ErrorFlag
would not be updated after the MD asymmetries had had
the blinding factor applied.

Just previous to this commit, using a global cut on an
MD channel would have the effect that the blinder failures
were applied globally, which is not what we had wanted to
do.

Compared against revision 3505 for run 12079:
Identical results are found if the MD and target position
cuts are deactivated; slightly different results are found
if the MD and target position cuts are active due to the
change in event selection.
Pattern by pattern, and event by event, the results agree
with revision 3505.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3558 b9700a98-9937-0410-b777-f318d6eb5bff