NOTE:  This is on branch release_2.00.
Increased the maximum number of leaves in the trees.

Modified the qwbatchsub.pl to support copying the split rootfiles
to the volitile disk.  Note that this temorarily *BREAKS* writing 
the root files to the silo.

Added support scripts for the official Run1Pass4 and Run2Pass1
analysis chains.
Regression is handled for now by just calling the linRegBlue
several times with different configuration files.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_2.00@3565 b9700a98-9937-0410-b777-f318d6eb5bff