Fixed compiler warnings in QwIntegratedRaster, QwBeamLine, QwBeamMod, QwLumi, QwMainCerenkovDetector (mostly unused variables, which were pruned).
In QwDBInterface there was function marked as inline that was only
defined in the cc file (and therefore certainly not inline).  Removed
the inline attribute.

Added option flag to suppress printing of error counter summary in QwParity.

Copy constructor for QwSubsystemArray (related to why Han had to add
a dummy pointer to get FillErrDB to work, I think).

Copy constructor (and Copy() method) for QwHelicity expanded.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3591 b9700a98-9937-0410-b777-f318d6eb5bff