Addition of a new tab in the Database GUI, QwDBGUI. This tab allows the quick plotting of popular graphs (MD/LUMI asyms, position differences, etc) for different pre-set time periods or a custom run range. You can currently select between plotting data from "Last 24 Hours", "Last Week", "Run to Date" (for Run II), or "Custom Run Range". The tab retains the basic run selection capabilities (run quality, good_for identity, target type, regression on/off) but is defaulted to production parity with good quality, regression off.
I'm still thinking about adding in some sort of "Post" button that could post results to the HCLog, similar to what Michael has in the AutoGUI. 

Please not that the plots chosen were simply those that I thought might be useful. I didn't pick any of the new BCM's because I wasn't sure what to show. I'd be happy to update it with anything that people want added.

Note that the structure of this GUI is such that automating certain plots isn't really possible. That is, one cannot set up a cron job to have certain plots made and then have them posted to some web directory: the GUI is too reliant on CLICKS.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3595 b9700a98-9937-0410-b777-f318d6eb5bff