I updated the DBGUI to view data for the different types of regression schemes we are currently using. There are currently 6 sets of regression schemes and each have different types of IVs.
(From Damon and Paul)
>>> reg vars 1        on
>>> diff_qwk_targetX
>>> diff_qwk_targetY
>>> diff_qwk_targetXSlope
>>> diff_qwk_targetYSlope
>>> diff_qwk_energy
>>>
>>> reg vars 2        on_5+1
>>> diff_qwk_targetX
>>> diff_qwk_targetY
>>> diff_qwk_targetXSlope
>>> diff_qwk_targetYSlope
>>> diff_qwk_energy
>>> asym_qwk_charge
>>>
>>> reg vars 3        on_set3
>>> diff_qwk_targetX
>>> diff_qwk_targetY
>>> diff_qwk_targetXSlope
>>> diff_qwk_targetYSlope
>>> diff_qwk_bpm3c12X
>>> asym_qwk_charge
>>>
>>> reg vars 4        on_set4
>>> diff_qwk_targetX
>>> diff_qwk_targetY
>>> diff_qwk_targetXSlope
>>> diff_qwk_targetYSlope
>>> diff_qwk_energy
>>> asym_qwk_bcm5
>>>
>>> reg vars 5        on_set5
>>> diff_bpm_9b_m_4X=diff_qwk_bpm3h09bX-diff_qwk_bpm3h04X
>>> diff_bpm_9b_m_4Y=diff_qwk_bpm3h09bY-diff_qwk_bpm3h04Y
>>> diff_bpm_9b_p_4X=diff_qwk_bpm3h09bX+diff_qwk_bpm3h04X
>>> diff_bpm_9b_p_4Y=diff_qwk_bpm3h09bY+diff_qwk_bpm3h04Y
>>> diff_qwk_bpm3c12X
>>>
>>> reg vars 6        on_set6
>>> diff_bpm_9b_m_4X=diff_qwk_bpm3h09bX-diff_qwk_bpm3h04X
>>> diff_bpm_9b_m_4Y=diff_qwk_bpm3h09bY-diff_qwk_bpm3h04Y
>>> diff_bpm_9b_p_4X=diff_qwk_bpm3h09bX+diff_qwk_bpm3h04X
>>> diff_bpm_9b_p_4Y=diff_qwk_bpm3h09bY+diff_qwk_bpm3h04Y
>>> diff_qwk_bpm3c12X
>>> asym_qwk_charge


I modified the drop down menu previously used to select regression on/off, to contain the names of these new schemes.

To get regressed/unregressed data and sensitivites, the data selection from the data base is done using following cuts:
To look at unregressed values we  use slope_correction='off' and slope_calculation='off'
To look at the regressed values we use slope_correction='on and slope_calculation='on_xx"
To look at sensitivities we  use slope_correction='off' and slope_calculation='on_xx'

As far as I can see from the preliminary checks I did, the queries work.
I do see some runlets with multiple entries but these are there with or without regression. I think this is something that is not related to a mistake in the query. I needs to be further investigated.

For qw_run1_pass3, we used slope_correction='off' and slope_calculation='off' to get the sensitivites. Therefore, to get the sensitivites from this database, users have to use regression 'off' from the regression type drop down menue.

I also implemented a print routine to print out which regression variables which are used in each regression scheme. These are supposed to be printed out when a user selects the Regression Type from the drop down menue. For some reason , I can't get them to print out once a plot has been drawn. They'll get printed out at the begining before you press the Submit button for the fist time. But after a plot has been drawn, the canvas remains empty. I'll look in to this later as it is not really really important at the moment.

Updated the axis labels of XXXSlopes to be Beam Angle (murad) instead of Beam Position (mm). Also updated the Effective Charge axis label to be Charge (arbitary units).
Modified the legend to display the fit results as exponentials. Previously they were displayed as floats with 3 decimal places but sometimes when both values and errors are smaller than that, the legends display zeros. Now they'll display the exponential values.

Removed GUIDatabase::DetectorVsMonitorPlot(). This is a function that was not used by anything. It was created as a part of an attempt to enable plotting vs monitors, but compared to the current status of the framework, it has become pretty much outdated and useless. 






git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3609 b9700a98-9937-0410-b777-f318d6eb5bff