Reverted the reversion of revision 3599 in revision 3625.
Errors gone; histogram AND console output now consistent with 3589.

Clarification:  When we push_back a subsystem in a subsystem array,
we automatically called the publishing function.  When we copy construct
a subsystem array from another subsystem array, the copy constructor first
copies the list of published variables and then pushes_back the contents
of the original subsystem array which again publishes the same variables.

Of course, on copies of subsystem arrays those published variables aren't
ever used... so the whole things didn't have any effect on the analysis.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3629 b9700a98-9937-0410-b777-f318d6eb5bff