Added two new plots to the realtime GUI. 1. asy_qwk_bcm1 2. yield_qwk_bpm3c12_EffectiveCharge.
The purpose of adding the effective charge of 3c12 is to give the shift crew and indication when the bpm starts to saturate. Based on what was observed in run 14153, the limits on the effective charge of 3c12 was set to be 90e3 and 190 e3 (+-50e3 around nominal values of 140e3). There is abox drawn on top of the histograms to indicate these limits. If the mean+rms value of the effective charge goes above 190e3, the error message "3c12BPM is saturating!" will be printed across the histograms in red. If the mean-rms goes below 90e3 the error message "3c12 BPM signal is too low!" will be printed across the histogram.

Since to add this new plot, I had to divide the canvas to 8 pads, I had a spare pad left and I decided to put the asymmetry on bcm1 on it. If anyone feels like there maybe another good variable that we want to keep and eye on instead of bcm1, please let me know and I can replace bcm1 histogram with that.

I tested the updated GUI on my machine. It seems to work fine. Need to test on cdaq cluseter.

Also added the bcm1 and 3c12 new histograms to the parity_realtime_hist_trim.in file.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3669 b9700a98-9937-0410-b777-f318d6eb5bff