The EPICS mapping file for the 2-pass beam for the start of Run2 was not committed at the beginning of Run2.
Because this file was missed, the Wien flip state for
the start of Run2 was interpreted without including the
effect of the extra precession in the accelerator.

For the Flip-right Wien state, the default configuration
was properly recognized, but the Flip-left state fell
into an indeterminate range.

We will not (and do not) introduce the extra precession
in this file, but correct the nominal launch angle to
report "normal" Wien state for Flip-right and "reversed"
for Flip-left.

NOTE:  When we go to 1-pass running in January, we will
       need to do something a little more clever to
       correctly account for the change in the precssion
       in the machine, since the use of the "launch angle"
       variable has been somewhat messed up through
       missing this file.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3684 b9700a98-9937-0410-b777-f318d6eb5bff