A few updates to the QwDBGUI as requested by M. Pitt and R. Subedi:
Added uslumi_sum to the main tab. Also changed the name of the summary tab Lumi entry to reflect that it is for dslumi's. Added a new button that shows uslumi_sum on the summary tab.

Added in the rest of the bcm double differences and changed how they are being retrieved from the database. Choosing regression off or on now properly displays unregressed/regressed BCm double differences.

Changed BCM double difference  y-axis label to be correct.

Tested on my personal copy and appears to be working fine. Buddhini or I will update the version on the cdaq cluster soon and make a log when it's ready.

git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3728 b9700a98-9937-0410-b777-f318d6eb5bff