Changed the EPICS class to report the IHWP polarity to the QwBlinder with an extra helicity flip based upon keywords found in the map file.
Thee are now three EPICS map files:  unversioned, which corresponds
to the Run 1 conditions; run range 13337-14699, which is the 2-pass
running in Run 2; and open run range 14700-, which will be the 1-pass
running for the remainder of Run 2.

These changes make it possible for us to keep the same blinding seed
for the 2-pass and 1-pass running of Run 2, if we choose to do so.

Functionality has been tested and compared against the release 2.00
branch for runs 12079 and 14023.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3739 b9700a98-9937-0410-b777-f318d6eb5bff