NOTE:  This is on branch release_2.00.
Merge of the changes from trunk revision 3739 into the release_2.00
branch.
This just picks up the new interaction between the EPICS class
and the blinder, to support keeping the same blinding seed for
the 2-pass and 1-pass beam where there is an extra spin flip
from precession.

Tested to produce the same results for runs 12079 and 14023.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_2.00@3743 b9700a98-9937-0410-b777-f318d6eb5bff