NOTE:  This is on branch release_2.00.
Merged the changes in Parity/prminput from trunk revision
range 3560:3743 into branch release_2.00.

This is to just pick up all the configuration file changes
that we've been using on the cdaq cluster into the 2.00
branch.

Confirmed to produce the same results as the official copy
of the engine for runs 12079 and 14023.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_2.00@3746 b9700a98-9937-0410-b777-f318d6eb5bff