Changed the routine names to make more sense as the feedback system is performing multiple feedbacks
Added target combo and BPM that belongs to the target
Published the target variables 
Started adding some functionality to calculate running averages for target variables.

To proceed from here it is best to test the feedback with the DAQ and probably with beam to make sure the important PITA feedback is working after many changes that I did.

Rakitha


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3761 b9700a98-9937-0410-b777-f318d6eb5bff