Removed the QwSIS3801_Module dependence in tracking subsystems that still had it.
For the innocent bystanders wondering what's happening with their subsystems, this
only puts the scaler channels directly in the subsystem instead of through a layer
of redirection that didn't turn out to be as useful as originally anticipated.

Checked with histos in tracking run 8670, but I am afraid that those histograms
weren't very representative.  Since it works for the scanner I am confident that
it works for these tracking subsystems too.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3773 b9700a98-9937-0410-b777-f318d6eb5bff