Commit changes to MakeSluglet and ppMakeSluglet with some updates and bug fixes. The C script now takes in the location of the runlet ROOTfiles as an argument. The bash script points it to the standard $QW_ROOTFILES directory. The new ppMakeSluglet_vol script does the same thing but looks at the volatile space instead. The output Sluglet ROOTfiles are written to $QW_ROOTFILES but this can be changed by editting the bash script(s).
What is this script? It is designed to be used as a post-processing script, run when all of the runlets of a run have been successfully analyzed by the qwparity analyzer. It will then create a single ROOTfile for the entire RUN (that is, not separated into segments) and call it Sluglet_RUNNUM.root. The file contains a heavily trimmed down list of leaves, as provided in the leaflist.txt file. 

This commit fixes a bug wherein the run segments would not be combined in the correct order. The solution was to replace the pair<Int_t,Int_t> _array_ that contains defines (run, runlet) pairs with a std::vector of the same format and use std::sort.

The script takes ~7 minutes to run on the ifarm looking at the volatile space and creates a file that is ~500 MB for the entire run. This is down from something on the order of 65 GB (that can't be right...) that I see looking at the individual ROOTfiles for the orginial runlets.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@3853 b9700a98-9937-0410-b777-f318d6eb5bff