This is just to update bcm2 calibration numbers for the run range 16057-16062 where bcm2 signal was plugged in into digital receiver box. I used a beam trip in runlets 16058.007 and 16058.008 (there was only one beam trip split into these two runlets) to calibrate bcm2 against bcm1, the calibration numbers being:
qwk_bcm2        ,-644.965,      0.00204878

For this, I had to close the existing open map file for the run range 15261-16056 and copy the same map file for another range 16063-.map, since the calibration numbers are the same. Though the run 16057 was a short transition run and I was also adjusting bcm2 gains in the digital box, the new calibration number may not be exact for this run.  


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4048 b9700a98-9937-0410-b777-f318d6eb5bff