Note: this is branch wdconinc.first_pass_regression!
Modification to get the double differences to work.  Don't know if they make sense
now, but they fill and it runs.  Used Paul's fix (private communication) and had to
fix a problem with the publication lists not being copied.  Added bcm5 and bcm6 to
the list of published variables for my favorite run 12079.  This will need to be done
for every hallc beamline map file.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.first_pass_regression@4071 b9700a98-9937-0410-b777-f318d6eb5bff