NOTE:  This is in branch wdconinc.first_pass_regression.
Correcting an incorrect change that entered in revision 3716;
this altered the order of operations by which the relative
stripline positions are turned into the absolute positions.

The way it should go is (and is now again):
	abs_position = (relative_position + center)/position_gain
as the centers and gains are determined from fitting the
relative position with the projected position from two BPMs
as the independent variable.

The trunk has maintained the proper order of operations.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.first_pass_regression@4114 b9700a98-9937-0410-b777-f318d6eb5bff