Note: this is on branch wdconinc.first_pass_regression.
Added a "publish by request" mechanism to the subsystem array:
if a requested variable is not found in the published list,
all of the subsystems will be asked if they contain that
variable, and to publish it if they do.

The reason for adding this mechanism is to allow us to more
easily change what variables are published based on what is
needed by the regression objects.

Right now, only the QwBeamLine subsystem has a complete
PublishByRequest function.  The variable names which can be
requested are similar to the names of the corresponding
variable in the MPS tree.  For instance, these variable
names would all be recognized by the QwBeamLine PublishByRequest
function:
	qwk_bcm1
	qwk_bpm3h09X
	qwk_bpm3h09_EffectiveCharge
	qwk_targetXSlope
The only variables that need to be explicitly published from
within the QwBeamLine class will be for variables whose published
name do not match their "tree name".

I have removed the explicit publication of the qwk_bcm# variables
from the two qweak_hallc_beamline.*.map files which had them
defined.

This has been tested on run 12079.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.first_pass_regression@4132 b9700a98-9937-0410-b777-f318d6eb5bff