Note:  this in in branch wdconinc.first_pass_regression.
Added the DB filling routine to the QwRegression class.
Modified the naming scheme for "aliases" in the QwRegression class;
they will not have a "reg_" prefix added to their names from the
regression map file.
Edited the regression map file to use alias names more consistent
with the usage in linRegBlue.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/wdconinc.first_pass_regression@4180 b9700a98-9937-0410-b777-f318d6eb5bff