Synchronize bursts at helicity pattern boundaries (finally).
Instead of ending a burst after a fixed number of helicity windows,
the bursts are now ended after a specified number of patterns. Most
of the burst support was already in the helicity pattern class, so
this was an easy fix.  Parity, Moller, and Compton executables have
been updated to compile, but only the Parity executable has reliable
bursts.  The remaining problem with the bursts is that they don't
contain any ErrorFlag info so it is hard to cut out beam trips.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4182 b9700a98-9937-0410-b777-f318d6eb5bff