Fixed the bug in the Database tab  which was plotting unregressed data instead of regressed data when asked to plot regressed data. It was a bug that was introduced in to the GUI when I did my commit on 27th February, 2012 (revision 4087).
The cause: 
With the commit on 27th, Feb. I hard coded the regression type to be "on" when bcm double differences are selected because it confuses people why we have to select regression 'on' to look at bcm double differences. I checked it was working and I check that the other selection types were working and thought all was well but it was not. It has confused the database tab somehow.

NOTE: This bug was only seen in the Qweak Database tab. It did not effect the Database Summary tab since the two tabs are generated from separate source files.  

.
The fix: Created a separate detector type for bcm double differences and enabled regression type to be 'on' and slope_calculation to be 'off' when that detector type is selected (no need to hard code anything). The regression type box is disabled for bcm double differences hoping not to confuse users. less options-> less confusion!
By adding bcm double differences as a separate array I can use the "Select()" functions of the combo boxes to preset the variables like regression type on the spot without having to hard code them later inside the MakeQuery() function after doing a hundred checks to figure out if a double difference is selected.

Other tweaks: 
- Added a new type in to the regression type box called the "raw(from rootfiles)" (if this is confusing, we can pick a suitable name later) to enable plotting values from the analyzer with standard cuts. This use the selection "slope_calculation = 'off' and slope_correction='off'. 
Till now,the GUI was only looking at the data that passed the cuts on the regression script.


- Switched the regression information text from TLatex to TPaveText. That makes the list readable.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4202 b9700a98-9937-0410-b777-f318d6eb5bff