Modified several beamline channel maps to reflect the move of the BCM5/6 electronics from the hall to the counting house on 03FEB2011.  I don't know why we didn't create a new run range when this work was done, but looking at some ROOT files from around this time, there is a discontinuity in the BCM5/6 response.
The run range file for 9494-12101 was split into
two run ranges:  9494-9686 and 9687-12101.

The VQWK channels corresponding to the cavity BPM
position channels were commented out for all run 
ranges starting with 9687-12101.
There may not be any useful information in the position
channels for earlier runs, but it is gone after the
electronics moved upstairs.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4225 b9700a98-9937-0410-b777-f318d6eb5bff