edetLasCyc.C and asymFit.C now does not do any plotting, all the plotting work is left on the shoulders of fileReadDraw.C
theoryAsym.C now skips the masked strips in the evaluation of asymmetry.
this was earlier causing an offset in the comparision between experimental and theoretical asymmetry

now the function of theoretical asymmetry is fit in one file while it is used in another, this is not good.
this may later lead to errors if I forget the way it is done, hopefuly it will get cleaner as I progress

trivial changes to other files



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4260 b9700a98-9937-0410-b777-f318d6eb5bff