Added in options to cut on recent additions to the good_for_GUI including "scanner_on" and "dis". Also added in the set9 option for plotting regressed results and sensitivities.
One can now plot, say, the MDallbars "sensitivity" to asym_uslumi_sum by choosing MD Sensitivities from the bottom of the Instrument Type box, choosing on_se9 as the regression type and wrt_uslumi_sum in the Regression IV box.

I've tested this on the DIS data of recent days and all appears to be working well. Please let me know if it is not working for you. I will update the version on the cdaq cluster shortly and HCLog it.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4352 b9700a98-9937-0410-b777-f318d6eb5bff