Modifications to the QwOptions class to be able to parse options, define a new config file, and then parse some more options.  Until now you had to stop adding config files when you wanted to parse the first file.
In practice, for the tracking analyzer, you can now specify a run-specific qwtracking.conf
file (see example) that will be used only for that run.  It contains the typical command line
options like octant setting, magnet current, etc.

This is not the best solution to the problem of which options should get preference when they
are specified multiple times (command line > config file, but which config files override
which others).  Best is not to rely on override behavior.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4399 b9700a98-9937-0410-b777-f318d6eb5bff