a significant change made in getEBeamLasCuts.C, the method now has two different cut-limits for hi and low of laser and beam cuts. earlier it was around one fixed fraction of the laserMax and beamMax that this cut used to happen
also, if the run is a low beam run (less than 65 uA), only the lower limit of the beamMax cut is used for declaring beamOn, 
Since we are charge normalizing, doing this is not very unsafe, 
though this way, I would be getting a bigger portion of beam-ramp-up data during which the beam may not be stable in BPM positions

the asymFit.C now also writes out a file with the polarization, polEr, chiSq, compton edge etc for all the active planes of this run.
This file can then later be used for a global review/analysis.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4416 b9700a98-9937-0410-b777-f318d6eb5bff