Added a new lumi channel map in which the USLumi5neg is reenabled; the USlumisum is still defined as only uslumi3 and uslumi7 though.
Modified the MD and lumi pedestal extraction script to attempt to
extract pedestals from all lumi channels; if a channel is not found
in the tree it is skipped, instead of causing the script to crash.

Determined pedestals for MD and lumi using the new version of the script.

Note that the new lumi map and pedestal analysis could be pushed
back earlier in time, once we know when uslumi5 started behaving again.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4420 b9700a98-9937-0410-b777-f318d6eb5bff