Software suite dedicated to retrieving data, specifically modulation data, from	the database, handling it, and making useful plots.
QwDBConnection: This is a wrapper class to handle Mysql++ DB connections in a format that make more sense to me.
QwDataContainer: This is a class to work with the data retrieved from the DB.
QwPlotHelper: Plotter class to make useful plot w/o the need to make a 10,000 line 'copy-paste' script...

All of this needs a lot of work (more thorough error check for instance) but I think this a good start. 


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4454 b9700a98-9937-0410-b777-f318d6eb5bff