have been making incremental changes towards proper noise correction. As of now, in this version, the noise correction is explicitly set to zero. Hence the laser-off background is simply charge normalized and subtracted without any consideration for electronic noise.
the getEBeamLasCuts.C has improved with an additional condition that not just the laser off 
but the laser on should also be atleast be 2second long for it to be recorded as a good laser on period.

it did not work on run 25509, due to some inconsistencies which hasn't been figured out yet, but
this version is tested to work on run 25484.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4480 b9700a98-9937-0410-b777-f318d6eb5bff