NOTE:  this is on branch release_2.02.
Now that the BMOD analysis includes cuts on the Device_Error_Code
of all the variables it is manipulating, we need to be careful to
exclude devices with a high failure rate.

Modifed the BMOD config file to exclude BCM5.
Also removed BCM7 and BCM8 since they don't exist for Run1.

Note that eventually we will need to modify BMOD to be able to use
run-number-dependent config files in order to support processing
both Run1 and Run2.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/release_2.02@4497 b9700a98-9937-0410-b777-f318d6eb5bff