changed the Residual code so that it now produces plots of the mean value of the redual and the error as the RMS and the RMS/sqrt(N) for each plane in each package.
Changed the run_page code so that the lay out is nicer and some plots are side by side so that they are easier to look at on the website

Created the QwHitsInfo.C code which graphs the QwHits.fRegion in a plot, QwHits.fPackage in a plot, QwHits.fDirection in a plot, and fQwHits.fElement in each region - and in region 2 and 3 by package number

Created OctatnDetermination.C which for each package in region 3 graphs the Maindetector signal vs trigger sicitnlator in each octant and in the center the octant number that is stated to by the config file as that package.  Also it graphs for region 3 just the main detector signal in each octant and in the center the octant number that is stated to by the config file as that package.  The lay out is the same as the octant layout in the main-dector bar of in the experiment. 




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4533 b9700a98-9937-0410-b777-f318d6eb5bff