the edetLasCyc.C would not call asymFit and fileReadDraw if the expAsym function does not execute happily
the evaluateAsym and expAsym is moved to an older version and updated with some of the new changes. This version has the laserCycle defined as a laser-off period followed by a laser-on period. 

comptonRunConstants has a few values changed to adapt for the noiseEstimation tests I am currently into.

getEBeamLasCuts now has only one cut value instead of the earlier 'upper cut' value and 'lower cut' value. That was deceiving. The idea of getEBeamLasCuts was to create
laser-off and beam-off regions and then we can use more strict cuts in the expAsym, while defining what I use as laserOn and BeamOn

Trivial changes to fileReadDraw.C
  


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4545 b9700a98-9937-0410-b777-f318d6eb5bff