Added QwEventBuffer::ReOpenStream() that will rewind the current stream without going to the next segment or run.
In QwTracking a (very fast) loop to the first EPICS event is now included.
It simply prints the QTOR DCCT setpoint, but this could be used to set all
kinds of stuff from EPICS without losing events (e.g. target status in the
blinder).  However, there doesn't seem to be a scalable way of implementing
this.  It will quickly become a list of "if this then that" statements.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4558 b9700a98-9937-0410-b777-f318d6eb5bff