Final support for changing the magnetic field on the fly using EPICS data (*). This now works properly, but is commented out to allow for more testing.
Also modified the way the QTOR current is read in a little bit.  Entering
the number in the beam_property files was never necessary, and a preferred
way is to use the regular option QwMagneticField.current (otherwise we can
get conflicts if there are two independent ways of changing this).
For run 15121, I changed this from beam_property to qwtracking.conf.

(*) This will be useful to analyze the tracking QTOR scans that were taken
as a signle run with the green monster.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4560 b9700a98-9937-0410-b777-f318d6eb5bff