Initial pass at updating the weight factors to be used when combining PMTs in the analyzer.
I add in two new mapfiles that correspond to the start (9594) and middle (9703) of Wien0. I also update the 9810 mapfile to use the correct scheme, as it looked as though that mapfile was using the w_i = 1/<Y_i> * 1/<RMS_i^2> scheme that was toyed with in the past.

I'm still looking through the weights with a fine-toothed comb but they all look relatively stable now with the changes mentioned above. There are plenty of mapfiles for runs prior to the start of Wien0 which I am leaving untouched for now.




git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4572 b9700a98-9937-0410-b777-f318d6eb5bff