If there is no magnetic field current in the configuration files, try to find it in the EPICS data stream.  Exit with error if there is nothing in the EPICS stream (or if you didn't consider enough events).  Otherwise keep track of the EPICS values and change the field if the current in EPICS changes.  Specified magnetic field current always has priority and stays constant.  This will allow us to analyze QTOR scans that were taken with the green monster.  Cleandata is in the Mps_Tree for tracking.
This commit made me realize that the system of units was not internally
consistent.  Tesla redefined in terms of SI units.  This breaks binary
field maps, so you will have to regenerate them if you use them.

Negative magnetic field polarities are not something the code was written
to handle, so (although this commit should work with negative currents)
there might be some work to be done there.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4582 b9700a98-9937-0410-b777-f318d6eb5bff