Commented out the cavity BPM X, Y, and phase channels corresponding to BCM5/6 from the beamline channel maps.
These map files were for runs before 9687, which is
when the recievers were moved to the 1st floor CHC.

By the best of my recollection, we never had very robust
data in these channels, and in general just have not looked
at them.  Those channels also do not have a database entry.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4690 b9700a98-9937-0410-b777-f318d6eb5bff