Modified both the MD and lumi eventcuts files to add a global cut on all PMTs and combined PMTs, with the exception of: 	qwk_isourc 	qwk_preamp 	qwk_cagesr
I note in passing that "qwk_isourc" appears to be failing the
VQWK saturation cut early in the run before we ran the battery
down.
If we want to add that to the global cuts, we'll need to do a little
cleanup on how the VQWK saturation level gets assigned.

On a test of 50k events for run 9605, these cuts allowed 11227
patterns out of 11237 which passed the expanded beamline cuts.
The original cut definitions passed 11243 patterns.



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@4693 b9700a98-9937-0410-b777-f318d6eb5bff